new file:   CharacterDesignGuide.txt 	new file:   include/characters.h 	modified:   include/gameObject.h 	modified:   include/player.h 	modified:   include/vectors.h 	modified:   src/game.c 	modified:   src/gameObject.c 	modified:   src/handleEvents.c 	modified:   src/main.c 	modified:   src/player.c 	modified:   src/textureManager.c 	modified:   src/vectors.c

+++ include/gameObject.h
@@ -19,7 +19,7 @@ typedef struct {
 	SDL_RendererFlip flip;
 } gameObject;
 
-int FUR_createObject( const char* fileName, int id, float x, float y, int width,
+int FUR_createObject( const char* fileName, int id, int x, int y, int width,
 		int height, int currentRow, int currentFrame, double angle,
 		int alpha, SDL_Renderer* pRenderer, SDL_RendererFlip flip );
